address reviews. improve testing
