+++ applications/source/debezium-source/src/test/java/org/springframework/cloud/stream/app/source/debezium/integration/DebeziumTestUtils.java
@@ -34,9 +34,12 @@ public final class DebeziumTestUtils {
 
 	public static final String DATABASE_NAME = "inventory";
 	public static final String BINDING_NAME = "debeziumSupplier-out-0";
-	public static final String IMAGE_TAG = "2.2.0.CR1";
+	public static final String IMAGE_TAG = "2.2.0.Final";
 	public static final String DEBEZIUM_EXAMPLE_MYSQL_IMAGE = "debezium/example-mysql:" + IMAGE_TAG;
 	public static final String DEBEZIUM_EXAMPLE_POSTGRES_IMAGE = "debezium/example-postgres:" + IMAGE_TAG;
+	public static final String DEBEZIUM_EXAMPLE_MONGODB_IMAGE = "debezium/example-mongodb:" + IMAGE_TAG;
+
+
 
 	private DebeziumTestUtils() {
