+++ functions/supplier/debezium-supplier/src/test/java/org/springframework/cloud/fn/supplier/debezium/custom/DebeziumCustomConsumerIntegrationTest.java
@@ -44,7 +44,7 @@ public class DebeziumCustomConsumerIntegrationTest {
 	private static final Log logger = LogFactory.getLog(DebeziumCustomConsumerIntegrationTest.class);
 
 	private static final String DATABASE_NAME = "inventory";
-	public static final String IMAGE_TAG = "2.2.0.CR1";
+	public static final String IMAGE_TAG = "2.2.0.Final";
 	public static final String DEBEZIUM_EXAMPLE_MYSQL_IMAGE = "debezium/example-mysql:" + IMAGE_TAG;
 
 	@TempDir
